@@ -68,81 +65,86 @@
     ],
     "*.md": [
       "npm run lint-markdown"
+    ],
+    "*.{js,jsx,css,scss,md,mdx,json}": [
+      "prettier --write"
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@babel/eslint-parser": "^7.12.1",
+    "@babel/core": "^7.13.8",
+    "@babel/eslint-parser": "^7.13.10",
     "@babel/plugin-proposal-class-properties": "^7.12.1",
-    "@babel/preset-env": "^7.12.7",
-    "@babel/preset-react": "^7.12.10",
+    "@babel/preset-env": "^7.13.9",
+    "@babel/preset-react": "^7.12.13",
     "@mdx-js/loader": "^1.6.21",
-    "@octokit/auth-action": "^1.3.0",
-    "@octokit/rest": "^18.0.5",
+    "@octokit/auth-action": "^1.3.2",
+    "@octokit/rest": "^18.3.4",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.2",
     "@svgr/webpack": "^5.5.0",
     "alex": "^5.1.0",
-    "autoprefixer": "^10.1.0",
+    "autoprefixer": "^10.2.5",
     "babel-loader": "^8.2.2",
-    "copy-webpack-plugin": "^7.0.0",
-    "cross-env": "^7.0.2",
-    "css-loader": "^5.0.1",
-    "css-minimizer-webpack-plugin": "^1.1.3",
-    "cypress": "^6.1.0",
-    "directory-tree": "^2.2.4",
+    "copy-webpack-plugin": "^8.0.0",
+    "css-loader": "^5.1.3",
+    "css-minimizer-webpack-plugin": "^1.3.0",
+    "cypress": "^6.6.0",
+    "directory-tree": "^2.2.7",
     "directory-tree-webpack-plugin": "^1.0.2",
     "duplexer": "^0.1.1",
-    "eslint": "^7.15.0",
+    "eslint": "^7.22.0",
+    "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-cypress": "^2.11.2",
-    "eslint-plugin-markdown": "^2.0.0-rc.0",
+    "eslint-plugin-markdown": "^2.0.0",
     "eslint-plugin-mdx": "^1.8.2",
-    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react": "^7.22.0",
+    "eslint-plugin-react-hooks": "^4.2.0",
     "front-matter": "^4.0.2",
-    "html-loader": "^1.3.0",
-    "html-webpack-plugin": "^5.0.0-beta.1",
+    "html-loader": "^2.1.2",
+    "html-webpack-plugin": "^5.3.0",
     "http-server": "^0.12.3",
-    "husky": "^4.3.5",
+    "husky": "^5.1.3",
     "hyperlink": "^4.6.0",
     "jest": "^26.6.3",
-    "lint-staged": "^10.5.3",
-    "lodash": "^4.17.19",
-    "markdownlint": "^0.22.0",
-    "markdownlint-cli": "^0.26.0",
-    "markdownlint-rule-emphasis-style": "^1.0.0",
-    "mini-css-extract-plugin": "^1.3.3",
+    "lint-staged": "^10.5.4",
+    "lodash": "^4.17.21",
+    "markdownlint": "^0.23.1",
+    "markdownlint-cli": "^0.27.1",
+    "mini-css-extract-plugin": "^1.3.9",
     "mkdirp": "^1.0.4",
     "modularscale-sass": "^3.0.3",
     "node-fetch": "^2.6.1",
     "npm-run-all": "^4.1.1",
-    "postcss": "^8.2.1",
-    "postcss-loader": "^4.1.0",
+    "postcss": "^8.2.8",
+    "postcss-loader": "^5.2.0",
+    "prettier": "^2.2.1",
     "react-refresh": "^0.9.0",
     "redirect-webpack-plugin": "^1.0.0",
     "remark": "^13.0.0",
     "remark-autolink-headings": "^6.0.1",
-    "remark-custom-blockquotes": "1.0.0",
+    "remark-emoji": "^2.1.0",
     "remark-extract-anchors": "1.1.1",
     "remark-frontmatter": "^3.0.0",
     "remark-gfm": "^1.0.0",
     "remark-html": "^13.0.1",
-    "remark-loader": "^2.0.0",
+    "remark-loader": "^3.0.0",
     "remark-refractor": "git://github.com/montogeek/remark-refractor.git",
     "remark-slug": "^6.0.0",
     "rimraf": "^3.0.2",
-    "sass": "^1.30.0",
-    "sass-loader": "^10.1.0",
-    "sirv-cli": "^1.0.10",
+    "sass": "^1.32.8",
+    "sass-loader": "^11.0.1",
+    "sirv-cli": "^1.0.11",
     "sitemap-static": "^0.4.2",
-    "start-server-and-test": "^1.11.0",
+    "start-server-and-test": "^1.12.0",
     "static-site-generator-webpack-plugin": "^3.4.1",
     "style-loader": "^2.0.0",
     "tap-spot": "^1.1.1",
-    "webpack": "^5.10.3",
-    "webpack-bundle-analyzer": "^4.2.0",
-    "webpack-cli": "^4.2.0",
+    "unist-util-visit": "^2.0.3",
+    "webpack": "^5.26.0",
+    "webpack-bundle-analyzer": "^4.4.0",
+    "webpack-cli": "^4.5.0",
     "webpack-dev-server": "^4.0.0-beta.0",
-    "webpack-merge": "^5.7.0",
-    "workbox-webpack-plugin": "^6.0.0"
+    "webpack-merge": "^5.7.3",
+    "workbox-webpack-plugin": "^6.1.2"
   },
   "dependencies": {
     "docsearch.js": "^2.5.2",
