@@ -2,14 +2,10 @@ import './Container.scss';
 import PropTypes from 'prop-types';
 Container.propTypes = {
   className: PropTypes.string,
-  children: PropTypes.node
+  children: PropTypes.node,
 };
-export default function Container (props = {}) {
+export default function Container(props = {}) {
   let { className = '' } = props;
 
-  return (
-    <div className={ `container ${className}` }>
-      { props.children }
-    </div>
-  );
+  return <div className={`container ${className}`}>{props.children}</div>;
 }
