@@ -6,18 +6,18 @@ import PropTypes from 'prop-types';
 
 export default class Contributors extends Component {
   static propTypes = {
-    contributors: PropTypes.array
-  }
+    contributors: PropTypes.array,
+  };
   state = {
-    inView: false
-  }
+    inView: false,
+  };
 
   handleInView = (inView) => {
     if (!inView) {
       return;
     }
     this.setState({ inView });
-  }
+  };
 
   render() {
     const { inView } = this.state;
@@ -28,23 +28,31 @@ export default class Contributors extends Component {
     }
 
     return (
-      <VisibilitySensor delayedCall
+      <VisibilitySensor
+        delayedCall
         partialVisibility
-        intervalDelay={ 300 }
-        onChange={ this.handleInView }>
+        intervalDelay={300}
+        onChange={this.handleInView}
+      >
         <div className="contributors">
           <div className="contributors__list">
-            {
-              contributors.map(contributor => (
-                <a key={ contributor }
-                  className="contributor"
-                  href={ `https://github.com/${contributor}` }>
-                  <img alt={ contributor }
-                    src={ inView ? `https://github.com/${contributor}.png?size=90` : SmallIcon } />
-                  <span className="contributor__name"> {contributor}</span>
-                </a>
-              ))
-            }
+            {contributors.map((contributor) => (
+              <a
+                key={contributor}
+                className="contributor"
+                href={`https://github.com/${contributor}`}
+              >
+                <img
+                  alt={contributor}
+                  src={
+                    inView
+                      ? `https://github.com/${contributor}.png?size=90`
+                      : SmallIcon
+                  }
+                />
+                <span className="contributor__name"> {contributor}</span>
+              </a>
+            ))}
           </div>
         </div>
       </VisibilitySensor>
