@@ -1,12 +1,8 @@
 import './Markdown.scss';
 import PropTypes from 'prop-types';
 
-const Markdown = props => (
-  <div className="markdown">
-    { props.children }
-  </div>
-);
+const Markdown = (props) => <div className="markdown">{props.children}</div>;
 Markdown.propTypes = {
-  children: PropTypes.node
+  children: PropTypes.node,
 };
 export default Markdown;
