@@ -11,6 +11,9 @@ import Dropdown from '../Dropdown/Dropdown';
 // Import helpers
 import isClient from '../../utilities/is-client';
 
+// Import constants
+import { THEME } from '../../constants/theme';
+
 // Load Styling
 import 'docsearch.js/dist/cdn/docsearch.css';
 import './Navigation.scss';
@@ -21,41 +24,47 @@ import TwitterIcon from '../../styles/icons/twitter.svg';
 import StackOverflowIcon from '../../styles/icons/stack-overflow.svg';
 
 const onSearch = () => {};
+const { DARK, LIGHT } = THEME;
 
 export default class Navigation extends Component {
   static propTypes = {
     pathname: PropTypes.string,
     links: PropTypes.array,
-    toggleSidebar: PropTypes.func
-  }
+    toggleSidebar: PropTypes.func,
+    theme: PropTypes.string,
+    switchTheme: PropTypes.func,
+  };
   render() {
-    let { pathname, links, toggleSidebar } = this.props;
+    const { pathname, links, toggleSidebar, theme, switchTheme } = this.props;
+    const themeSwitcher = () => switchTheme(theme === DARK ? LIGHT : DARK);
 
     return (
       <Banner
         onSearch={onSearch}
         blockName="navigation"
-        logo={ <Logo light={ true } /> }
-        url={ pathname }
+        logo={<Logo light={true} />}
+        url={pathname}
         items={[
           ...links,
           {
             title: 'GitHub Repository',
             url: 'https://github.com/webpack/webpack',
             className: 'navigation__item--icon',
-            content: <GithubIcon aria-hidden="true" fill="#fff" width={16} />
+            content: <GithubIcon aria-hidden="true" fill="#fff" width={16} />,
           },
           {
             title: 'webpack on Twitter',
             url: 'https://twitter.com/webpack',
             className: 'navigation__item--icon',
-            content: <TwitterIcon aria-hidden="true" fill="#fff" width={16} />
+            content: <TwitterIcon aria-hidden="true" fill="#fff" width={16} />,
           },
           {
             title: 'webpack on Stack Overflow',
             url: 'https://stackoverflow.com/questions/tagged/webpack',
             className: 'navigation__item--icon',
-            content: <StackOverflowIcon aria-hidden="true" fill="#fff" width={16} />
+            content: (
+              <StackOverflowIcon aria-hidden="true" fill="#fff" width={16} />
+            ),
           },
           {
             className: 'navigation__item--icon',
@@ -63,14 +72,38 @@ export default class Navigation extends Component {
               <Dropdown
                 className="navigation__languages"
                 items={[
-                  { title: 'English', url: 'https://webpack.js.org/' },
-                  { lang: 'zh', title: '中文', url: 'https://webpack.docschina.org/' }
-                ]} />
-            )
-          }
+                  {
+                    title: 'English',
+                    url: 'https://webpack.js.org/',
+                  },
+                  {
+                    lang: 'zh',
+                    title: '中文',
+                    url: 'https://webpack.docschina.org/',
+                  },
+                ]}
+              />
+            ),
+          },
+          {
+            className: 'navigation__item--icon',
+            content: (
+              <button
+                style={{
+                  background: 'transparent',
+                  border: 'none',
+                  cursor: 'pointer',
+                }}
+                onClick={themeSwitcher}
+              >
+                {theme === DARK ? '🌙' : '☀️'}
+              </button>
+            ),
+          },
         ]}
-        link={ Link }
-        onMenuClick={ toggleSidebar } />
+        link={Link}
+        onMenuClick={toggleSidebar}
+      />
     );
   }
 
@@ -81,7 +114,7 @@ export default class Navigation extends Component {
       DocSearch({
         apiKey: 'fac401d1a5f68bc41f01fb6261661490',
         indexName: 'webpack-js-org',
-        inputSelector: '.navigation-search__input'
+        inputSelector: '.navigation-search__input',
       });
     }
   }
