@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 // Styles
 import './PageNotFound.scss';
 
-export default function PageNotFound () {
+export default function PageNotFound() {
   return (
     <div className="page markdown">
       <h1>Page Not Found</h1>
