@@ -1,7 +1,7 @@
 import './Placeholder.scss';
 
 // Placeholder string
-const PlaceholderString = () => (`
+const PlaceholderString = () => `
   <div class="placeholder">
     <p class="placeholder__medium">&nbsp;</p>
     <p class="placeholder__large">&nbsp;</p>
@@ -19,7 +19,7 @@ const PlaceholderString = () => (`
     <p class="placeholder__small">&nbsp;</p>
     <p class="placeholder__medium">&nbsp;</p>
   </div>
-`);
+`;
 
 function PlaceholderComponent() {
   return (
@@ -31,7 +31,4 @@ function PlaceholderComponent() {
   );
 }
 
-export {
-  PlaceholderString,
-  PlaceholderComponent
-};
+export { PlaceholderString, PlaceholderComponent };
