@@ -11,12 +11,12 @@ const PRINTABLE_SECTIONS = [
   'guides',
   'loaders',
   'migrate',
-  'plugins'
+  'plugins',
 ];
 Print.propTypes = {
-  url: PropTypes.string
+  url: PropTypes.string,
 };
-export default function Print (props) {
+export default function Print(props) {
   const { url } = props;
   const printUrl = _printPageUrlFromUrl(url);
 
@@ -27,13 +27,20 @@ export default function Print (props) {
 
   return (
     <div className="sidebar-item sidebar-item--disabled`">
-      <BarIcon className="sidebar-item__toggle" width={15} height={17} fill="#175d96" />
-      <a className="sidebar-item__title sidebar-link__print"
+      <BarIcon
+        className="sidebar-item__toggle"
+        width={15}
+        height={17}
+        fill="#175d96"
+      />
+      <a
+        className="sidebar-item__title sidebar-link__print"
         href={printUrl}
         rel="nofollow noopener noreferrer"
         alt="Print"
         title="Print"
-        target="_blank">
+        target="_blank"
+      >
         Print Section
         <img src={icon} />
       </a>
@@ -43,5 +50,7 @@ export default function Print (props) {
 
 function _printPageUrlFromUrl(urlRaw) {
   let urlSplit = urlRaw.split('/');
-  return PRINTABLE_SECTIONS.includes(urlSplit[1]) ? `/${urlSplit[1]}/printable/` : false;
+  return PRINTABLE_SECTIONS.includes(urlSplit[1])
+    ? `/${urlSplit[1]}/printable/`
+    : false;
 }
