@@ -6,8 +6,6 @@ window.matchMedia('print').addListener(function(mql) {
 });
 window.print();
 `;
-export default function PrintScript () {
-  return (
-    <script dangerouslySetInnerHTML={{ __html: printScript }}></script>
-  );
+export default function PrintScript() {
+  return <script dangerouslySetInnerHTML={{ __html: printScript }}></script>;
 }
