@@ -6,17 +6,17 @@ import PropTypes from 'prop-types';
 
 // TODO: Check to make sure all pages are shown and properly sorted
 export default class SidebarMobile extends Component {
-  _container = null
-  _initialTouchPosition = {}
-  _lastTouchPosition = {}
+  _container = null;
+  _initialTouchPosition = {};
+  _lastTouchPosition = {};
 
   static propTypes = {
     isOpen: PropTypes.bool,
     toggle: PropTypes.func,
-    sections: PropTypes.array
-  }
+    sections: PropTypes.array,
+  };
   render() {
-    let {isOpen, toggle} = this.props;
+    let { isOpen, toggle } = this.props;
     let openMod = isOpen ? ' sidebar-mobile--visible' : '';
 
     this._toggleBodyListener(isOpen);
@@ -24,23 +24,25 @@ export default class SidebarMobile extends Component {
     return (
       <nav
         className={`sidebar-mobile${openMod}`}
-        ref={ref => this._container = ref}
+        ref={(ref) => (this._container = ref)}
         onTouchStart={this._handleTouchStart}
         onTouchMove={this._handleTouchMove}
-        onTouchEnd={this._handleTouchEnd}>
-
+        onTouchEnd={this._handleTouchEnd}
+      >
         <div
           className="sidebar-mobile__toggle"
           onTouchStart={this._handleTouchStart}
           onTouchMove={this._handleOpenerTouchMove}
-          onTouchEnd={this._handleTouchEnd} />
+          onTouchEnd={this._handleTouchEnd}
+        />
 
         <div className="sidebar-mobile__content">
           <span
             role="button"
             className="sidebar-mobile__close"
-            onClick={toggle.bind(null, false)}>
-              <CloseIcon fill="#fff" width={20} />
+            onClick={toggle.bind(null, false)}
+          >
+            <CloseIcon fill="#fff" width={20} />
           </span>
 
           {this._getSections()}
@@ -49,11 +51,11 @@ export default class SidebarMobile extends Component {
     );
   }
 
-  _toggleBodyListener = add => {
+  _toggleBodyListener = (add) => {
     let actionName = add ? 'addEventListener' : 'removeEventListener';
     window[actionName]('touchstart', this._handleBodyClick);
     window[actionName]('mousedown', this._handleBodyClick);
-  }
+  };
 
   /**
    * Get markup for each section
@@ -67,18 +69,22 @@ export default class SidebarMobile extends Component {
       pathname = window.location.pathname;
     }
 
-    return this.props.sections.map(section => {
+    return this.props.sections.map((section) => {
       let active = section.url !== '/' && pathname.startsWith(section.url);
 
       return (
         <div
-          className={`sidebar-mobile__section ${active ? 'sidebar-mobile__section--active' : ''}`}
-          key={section.url}>
+          className={`sidebar-mobile__section ${
+            active ? 'sidebar-mobile__section--active' : ''
+          }`}
+          key={section.url}
+        >
           <Link
             className="sidebar-mobile__section-header"
             key={section.url}
             to={section.url}
-            onClick={this.props.toggle.bind(null, false)}>
+            onClick={this.props.toggle.bind(null, false)}
+          >
             <h3>{section.title || section.url}</h3>
           </Link>
 
@@ -101,16 +107,19 @@ export default class SidebarMobile extends Component {
       pathname = window.location.pathname;
     }
 
-    return pages.map(page => {
+    return pages.map((page) => {
       let url = `${page.url}`,
         active = pathname === url;
 
       return (
         <Link
           key={url}
-          className={`sidebar-mobile__page sidebar-mobile__section-child ${active ? 'sidebar-mobile__page--active' : ''}`}
+          className={`sidebar-mobile__page sidebar-mobile__section-child ${
+            active ? 'sidebar-mobile__page--active' : ''
+          }`}
           to={url}
-          onClick={this.props.toggle.bind(null, false)}>
+          onClick={this.props.toggle.bind(null, false)}
+        >
           {page.title}
         </Link>
       );
@@ -122,22 +131,22 @@ export default class SidebarMobile extends Component {
    *
    * @param {object} e - Native click event
    */
-  _handleBodyClick = e => {
-    const {isOpen, toggle} = this.props;
+  _handleBodyClick = (e) => {
+    const { isOpen, toggle } = this.props;
     if (isOpen && !this._container.contains(e.target)) {
       toggle(false);
     }
-  }
+  };
 
-  _handleTouchStart = e => {
+  _handleTouchStart = (e) => {
     this._initialTouchPosition.x = e.touches[0].pageX;
     this._initialTouchPosition.y = e.touches[0].pageY;
 
     // For instant transform along with the touch
     this._container.classList.add('no-delay');
-  }
+  };
 
-  _handleTouchMove = e => {
+  _handleTouchMove = (e) => {
     let xDiff = this._initialTouchPosition.x - e.touches[0].pageX;
     let yDiff = this._initialTouchPosition.y - e.touches[0].pageY;
     let factor = Math.abs(yDiff / xDiff);
@@ -149,9 +158,9 @@ export default class SidebarMobile extends Component {
       this._lastTouchPosition.x = e.touches[0].pageX;
       this._lastTouchPosition.y = e.touches[0].pageY;
     }
-  }
+  };
 
-  _handleOpenerTouchMove = e => {
+  _handleOpenerTouchMove = (e) => {
     let xDiff = e.touches[0].pageX - this._initialTouchPosition.x;
     let yDiff = this._initialTouchPosition.y - e.touches[0].pageY;
     let factor = Math.abs(yDiff / xDiff);
@@ -163,10 +172,10 @@ export default class SidebarMobile extends Component {
       this._lastTouchPosition.x = e.touches[0].pageX;
       this._lastTouchPosition.y = e.touches[0].pageY;
     }
-  }
+  };
 
-  _handleTouchEnd = e => {
-    const {isOpen} = this.props;
+  _handleTouchEnd = (e) => {
+    const { isOpen } = this.props;
     const threshold = 20;
 
     // Free up all the inline styling
@@ -174,13 +183,19 @@ export default class SidebarMobile extends Component {
     this._container.style.transform = '';
 
     // are we open?
-    if (isOpen && this._initialTouchPosition.x - this._lastTouchPosition.x > threshold) {
+    if (
+      isOpen &&
+      this._initialTouchPosition.x - this._lastTouchPosition.x > threshold
+    ) {
       // this is in top level nav callback
       this.props.toggle(false);
-    } else if (!isOpen && this._lastTouchPosition.x - this._initialTouchPosition.x > threshold) {
+    } else if (
+      !isOpen &&
+      this._lastTouchPosition.x - this._initialTouchPosition.x > threshold
+    ) {
       this.props.toggle(true);
       e.preventDefault();
       e.stopPropagation();
     }
-  }
+  };
 }
