@@ -7,11 +7,11 @@ import './Vote.scss';
 
 export default class Vote extends Component {
   state = {
-    VoteApp: null
-  }
+    VoteApp: null,
+  };
   componentDidMount() {
     if (window && !window.__ssgrun) {
-      import('webpack.vote').then(VoteApp => {
+      import('webpack.vote').then((VoteApp) => {
         this.setState({ VoteApp: VoteApp.default || VoteApp });
       });
     }
@@ -20,7 +20,7 @@ export default class Vote extends Component {
     const { VoteApp } = this.state;
     return (
       <Container className="vote markdown">
-        { VoteApp ? <VoteApp development={ false } /> : null }
+        {VoteApp ? <VoteApp development={false} /> : null}
       </Container>
     );
   }
