@@ -26,20 +26,26 @@ function buildContentTree(source, output) {
     return console.error('build-content-tree: you must provide a source path');
   }
   if (!output) {
-    return console.error('build-content-tree: you must provide a output file name');
+    return console.error(
+      'build-content-tree: you must provide a output file name'
+    );
   }
 
   let content = directoryTree(source, { extensions: /\.(md|mdx)/ });
 
   content = restructure(content, {
-    dir: source
+    dir: source,
   });
 
-  fs.writeFileSync(path.resolve(output), JSON.stringify(content, 2), (error) => {
-    if (error) {
-      console.log('scripts/build-content-tree', error);
-    } else {
-      console.log('Successfully built content tree file at ' + output);
+  fs.writeFileSync(
+    path.resolve(output),
+    JSON.stringify(content, 2),
+    (error) => {
+      if (error) {
+        console.log('scripts/build-content-tree', error);
+      } else {
+        console.log('Successfully built content tree file at ' + output);
+      }
     }
-  });
+  );
 }
