@@ -1,28 +1,78 @@
-import { precacheAndRoute, matchPrecache } from 'workbox-precaching';
+import { cacheNames } from 'workbox-core';
 import { registerRoute } from 'workbox-routing';
 import { CacheableResponsePlugin } from 'workbox-cacheable-response';
-import { CacheFirst, NetworkOnly } from 'workbox-strategies';
+import {
+  NetworkFirst,
+  StaleWhileRevalidate,
+  NetworkOnly,
+} from 'workbox-strategies';
 import { ExpirationPlugin } from 'workbox-expiration';
-import { setDefaultHandler, setCatchHandler } from 'workbox-routing';
+import { setCatchHandler, setDefaultHandler } from 'workbox-routing';
 import ssgManifest from '../dist/ssg-manifest.json';
 
-// Precache assets built with webpack
-precacheAndRoute(self.__WB_MANIFEST);
+const cacheName = cacheNames.runtime;
 
-precacheAndRoute(ssgManifest);
-
-// Precache manifest.json as ssgManifest couldn't catch it
-precacheAndRoute([
+const manifest = self.__WB_MANIFEST;
+const otherManifest = [
   {
     url: '/manifest.json',
-    revision: '1', // manually update needed when content changed
   },
-]);
+];
+const manifestURLs = [...manifest, ...ssgManifest, ...otherManifest].map(
+  (entry) => {
+    const url = new URL(entry.url, self.location);
+    return url.href;
+  }
+);
+self.addEventListener('install', (event) => {
+  event.waitUntil(
+    caches.open(cacheName).then((cache) => {
+      return cache.addAll(manifestURLs);
+    })
+  );
+});
+
+self.addEventListener('activate', (event) => {
+  // - [x] clean up workbox precached data
+  // TODO to be removed after maybe two months?
+  // i.e., 2021-03-23
+  event.waitUntil(
+    caches.delete(cacheNames.precache).then((result) => {
+      if (result) {
+        console.log('Precached data removed');
+      } else {
+        console.log('No precache found');
+      }
+    })
+  );
+});
+self.addEventListener('activate', (event) => {
+  // - [x] clean up outdated runtime cache
+  event.waitUntil(
+    caches.open(cacheName).then((cache) => {
+      // clean up those who are not listed in manifestURLs
+      cache.keys().then((keys) => {
+        keys.forEach((request) => {
+          if (!manifestURLs.includes(request.url)) {
+            cache.delete(request);
+          }
+        });
+      });
+    })
+  );
+});
+
+registerRoute(
+  ({ url }) => manifestURLs.includes(url.href),
+  new NetworkFirst({
+    cacheName,
+  })
+);
 
 // Cache Google Fonts
 registerRoute(
   /https:\/\/fonts\.gstatic\.com/,
-  new CacheFirst({
+  new StaleWhileRevalidate({
     cacheName: 'google-fonts-cache',
     plugins: [
       // Ensure that only requests that result in a 200 status are cached
@@ -39,11 +89,22 @@ registerRoute(
 );
 
 setDefaultHandler(new NetworkOnly());
+
+// fallback to app-shell for document request
 setCatchHandler(({ event }) => {
   switch (event.request.destination) {
     case 'document':
-      return matchPrecache('/app-shell/index.html');
+      return caches.match('/app-shell/index.html');
     default:
       return Response.error();
+  }
+});
+
+// TODO remove this in the future
+// as we are using NetworkFirst strategy now
+// TODO remove NotifyBox as well
+self.addEventListener('message', (event) => {
+  if (event.data && event.data.type === 'SKIP_WAITING') {
+    self.skipWaiting();
   }
 });
