@@ -15,7 +15,8 @@ const excludedLoaders = [
   'webpack-contrib/yaml-frontmatter-loader',
   'webpack-contrib/restyle-loader',
   'webpack-contrib/gzip-loader',
-  'webpack-contrib/cache-loader'
+  'webpack-contrib/cache-loader',
+  'webpack-contrib/json5-loader',
 ];
 const excludedPlugins = [
   'webpack-contrib/component-webpack-plugin',
@@ -23,10 +24,10 @@ const excludedPlugins = [
   'webpack-contrib/i18n-webpack-plugin',
   'webpack-contrib/babel-minify-webpack-plugin',
   'webpack-contrib/uglifyjs-webpack-plugin',
-  'webpack-contrib/zopfli-webpack-plugin'
+  'webpack-contrib/zopfli-webpack-plugin',
 ];
 
 module.exports = {
   excludedLoaders,
-  excludedPlugins
+  excludedPlugins,
 };
