@@ -2,8 +2,10 @@
 const fs = require('fs');
 
 module.exports = function (fileSuffixes = []) {
-    const filesInDist = fs.readdirSync('./dist');
-    return fileSuffixes.length
-        ? filesInDist.filter((file) => fileSuffixes.find(suffix => file.endsWith(suffix)))
-        : [];
+  const filesInDist = fs.readdirSync('./dist');
+  return fileSuffixes.length
+    ? filesInDist.filter((file) =>
+        fileSuffixes.find((suffix) => file.endsWith(suffix))
+      )
+    : [];
 };
