@@ -2,10 +2,10 @@ module.exports = function yamlHeadmatter(fields) {
   let ret = '---\n';
 
   Object.keys(fields).map((field) => {
-    ret += `${ field }: ${ fields[field] }\n`;
+    ret += `${field}: ${fields[field]}\n`;
   });
 
-  ret = `${ ret }---\n`;
+  ret = `${ret}---\n`;
 
   return ret;
 };
