@@ -2,36 +2,32 @@ const path = require('path');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const webpack = require('webpack');
 const h = require('hastscript');
-// FIXME
-// why updating source code in responsive-table.js won't invalidate the webpack cache???
-// see https://github.com/webpack/changelog-v5/blob/master/guides/persistent-caching.md#build-dependencies
-const responsiveTable = require('./src/plugins/responsive-table/responsive-table.js');
+const remarkResponsiveTable = require('./src/remark-plugins/remark-responsive-table/remark-responsive-table.js');
 const mdPlugins = [
   require('remark-gfm'),
   require('remark-slug'),
-  responsiveTable,
+  remarkResponsiveTable,
+  require('remark-emoji'),
   [
-    require('remark-custom-blockquotes'),
+    require('./src/remark-plugins/remark-custom-asides/index.js'),
     {
       mapping: {
         'T>': 'tip',
         'W>': 'warning',
-        '?>': 'todo'
-      }
-    }
+        '?>': 'todo',
+      },
+    },
   ],
   [
     require('remark-autolink-headings'),
     {
       behavior: 'append',
       content() {
-        return [
-          h('span.header-link')
-        ];
-      }
-    }
+        return [h('span.header-link')];
+      },
+    },
   ],
-  require('remark-refractor')
+  require('remark-refractor'),
 ];
 
 module.exports = () => ({
@@ -41,16 +37,20 @@ module.exports = () => ({
     buildDependencies: {
       config: [__filename],
     },
-    cacheDirectory: path.resolve(__dirname, '.cache/webpack')
+    cacheDirectory: path.resolve(__dirname, '.cache/webpack'),
   },
   resolve: {
     extensions: ['.js', '.jsx', '.scss'],
     fallback: {
-      path: require.resolve('path-browserify')
-    }
+      path: require.resolve('path-browserify'),
+    },
   },
   module: {
     rules: [
+      {
+        test: /react-spring/,
+        sideEffects: true,
+      },
       {
         test: /\.mdx$/,
         use: [
@@ -58,41 +58,35 @@ module.exports = () => ({
           {
             loader: '@mdx-js/loader',
             options: {
-              remarkPlugins: mdPlugins
-            }
-          }
-        ]
+              remarkPlugins: mdPlugins,
+            },
+          },
+        ],
       },
       {
         test: /\.md$/,
         use: [
           {
-            loader: 'html-loader'
+            loader: 'html-loader',
           },
           {
             loader: 'remark-loader',
             options: {
               remarkOptions: {
-                plugins: [...mdPlugins, require('remark-html')]
-              }
-            }
-          }
-        ]
+                plugins: [...mdPlugins, require('remark-html')],
+              },
+            },
+          },
+        ],
       },
       {
         test: /\.jsx?$/,
         exclude: /node_modules/,
-        use: [
-          'babel-loader'
-        ]
+        use: ['babel-loader'],
       },
       {
         test: /\.css$/,
-        use: [
-          MiniCssExtractPlugin.loader,
-          'css-loader',
-          'postcss-loader'
-        ]
+        use: [MiniCssExtractPlugin.loader, 'css-loader', 'postcss-loader'],
       },
       {
         test: /\.scss$/,
@@ -104,53 +98,53 @@ module.exports = () => ({
             loader: 'sass-loader',
             options: {
               sassOptions: {
-                includePaths: [ path.join('./src/styles/partials') ]
-              }
-            }
-          }
-        ]
+                includePaths: [path.join('./src/styles/partials')],
+              },
+            },
+          },
+        ],
       },
       {
         test: /\.woff2?$/,
         type: 'asset/resource',
         generator: {
-          filename: 'font/[name].[hash][ext][query]'
-        }
+          filename: 'font/[name].[hash][ext][query]',
+        },
       },
       {
         test: /\.(jpg|jpeg|png|ico)$/i,
         type: 'asset/resource',
         generator: {
-          filename: '[name].[hash][ext][query]'
-        }
+          filename: '[name].[hash][ext][query]',
+        },
       },
       {
         test: /\.svg$/i,
         type: 'asset/resource',
         exclude: [path.resolve(__dirname, 'src/styles/icons')],
         generator: {
-          filename: '[name].[hash][ext][query]'
-        }
+          filename: '[name].[hash][ext][query]',
+        },
       },
       {
         test: /\.svg$/i,
         use: ['@svgr/webpack'],
-        include: [path.resolve(__dirname, 'src/styles/icons')]
-      }
-    ]
+        include: [path.resolve(__dirname, 'src/styles/icons')],
+      },
+    ],
   },
   plugins: [
     new MiniCssExtractPlugin({
-      filename: '[name].[contenthash].css'
+      filename: '[name].[contenthash].css',
     }),
     new webpack.DefinePlugin({
       // https://github.com/algolia/algoliasearch-client-javascript/issues/764
-      'process.env.RESET_APP_DATA_TIMER': JSON.stringify('') // fix for algoliasearch
-    })
+      'process.env.RESET_APP_DATA_TIMER': JSON.stringify(''), // fix for algoliasearch
+    }),
   ],
   output: {
     path: path.resolve(__dirname, './dist'),
     publicPath: '/',
-    filename: '[name].bundle.js'
-  }
+    filename: '[name].bundle.js',
+  },
 });
